refactor(phpdoc): move PhpDoc and PhpDocTag from php-ast to php-parser
These types are the output of phpdoc::parse() and are never embedded
in the AST tree. Keeping them in php-ast was misleading — the AST crate
should contain only types that appear in the parsed syntax tree.

Moving them to php-parser (alongside phpdoc.rs where they are defined
and used) correctly reflects their role as parser output, not AST nodes.
No public API change: consumers access them via php_rs_parser::phpdoc.
