+++ crates/php-ast/src/lib.rs
@@ -4,6 +4,3 @@ pub mod visitor;
 
 pub use ast::*;
 pub use span::Span;
-
-// Re-export PHPDoc types at crate root for convenience
-pub use ast::{PhpDoc, PhpDocTag};
